Make revisions on the skill-related notes
In this case, it's mostly about the additional perspectives on how
learning works from the act of managing your information through various
note-taking methods.

I also restructured the note on org-babel and moved as its own note on
the hierarchical notebook. I think summarizing a tool and giving my own
comments about it is a nicer way of describing it. Plus, I can freely
link between any other types of note so I figured it would be better.

That said, I should be picky on how to make org-roam entries. And also
org-roam v2 is better, after all. :)