Create entries on potential traps on learning
Ehh... Why I didn't think of those things but here we are. They are
going to stay for a little while.

Also, there's a slight shift in how I want use my wiki now. Instead of
trying to craft every note into a set of evergreen notes. While Andy
Matuschak's notes are very practical, it also focuses too much
on crafting evergreen notes (a bit too much). Or maybe that was just my
impression? Either way, I'll let some of my thoughts to drift if it able
to link somewhere in the notes. I think I'm getting the point of
Luhmann's original use of Zettelkasten as a research partner.