Add notes on Emacs and creativity
Not much connection between the two (yet) but just something I keep in
mind for the previous days. For the creativity, I focused first on what
creativity is and how I think of it. If ever I have a random realization
then I'll be sure to continue this note or something.