Make notes on creativity and ideas
With updates on how it relates on note-taking (among other things).
Since I have note-taking notes, that's the only thing to relate to right
now but there is a lot of untouched thoughts on other things such as
programming and illustration generally how inspirations and innovations
came from already existing things.

That's how successors are generally made but that goes for another
writing session.