Start journalling for learning how to draw
This is a note format I'm going to experiment. It is a very nice format
for culminating the journey of learning something without making too
much nodes. Thankfully, org-roam is nice that it lets you choose how to
structure your notes with nodes as the atomic unit and it is working so far.