Create the recipes format
The recipes could be made into a separate repo of its own but it is my
notebook and I'm going to tkae advantage of it. The committed recipes
should give an idea of how it's going to be structured with splitting of
recipes whenever possible.

Take note that you shouldn't make recipes with perfect structure out of
the blue. Just like you manage your system configs, just put it all in
one recipe then split it when there's another recipe with a similar step
that results in a recipe.